rename files to match what the object inside it represents
