+++ src/background/stores/base-store-impl.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { BaseStore } from '../../common/base-store';
 import { Store } from '../../common/flux/store';
-import { BaseStore } from '../../common/istore';
 import { StoreNames } from '../../common/stores/store-names';
 
 export abstract class BaseStoreImpl<TState> extends Store implements BaseStore<TState> {
