+++ src/background/stores/details-view-store.ts
@@ -9,7 +9,7 @@ import { ContentActions } from '../actions/content-actions';
 import { DetailsViewActions } from '../actions/details-view-actions';
 import { ScopingActions } from '../actions/scoping-actions';
 import { PreviewFeaturesActions } from './../actions/preview-features-actions';
-import { BaseStoreImpl } from './base-store';
+import { BaseStoreImpl } from './base-store-impl';
 
 export class DetailsViewStore extends BaseStoreImpl<IDetailsViewData> {
     constructor(
