+++ src/background/stores/istore-hub.d.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { BaseStore } from '../../common/istore';
+import { BaseStore } from '../../common/base-store';
 import { StoreType } from '../../common/types/store-type';
 
 export interface StoreHub {
