+++ src/common/base-store.ts
