+++ src/injected/layered-details-dialog-component.tsx
@@ -4,8 +4,8 @@ import { getRTL } from '@uifabric/utilities';
 import { LayerHost } from 'office-ui-fabric-react';
 import * as React from 'react';
 
+import { BaseStore } from '../common/base-store';
 import { FeatureFlags } from '../common/feature-flags';
-import { BaseStore } from '../common/istore';
 import { DevToolActionMessageCreator } from '../common/message-creators/dev-tool-action-message-creator';
 import { DevToolState } from '../common/types/store-data/idev-tool-state';
 import { UserConfigurationStoreData } from '../common/types/store-data/user-configuration-store';
