+++ src/tests/unit/common/assessment-store-tester.ts
@@ -3,7 +3,7 @@
 import { IMock, It, Times } from 'typemoq';
 
 import { IndexedDBDataKeys } from '../../../background/IndexedDBDataKeys';
-import { BaseStoreImpl } from '../../../background/stores/base-store';
+import { BaseStoreImpl } from '../../../background/stores/base-store-impl';
 import { IndexedDBAPI } from '../../../common/indexedDB/indexedDB';
 import { DefaultConstructor } from '../../../common/types/idefault-constructor';
 import { StoreTester } from './store-tester';
