+++ src/tests/unit/tests/background/assessment-scan-policy-runner.test.ts
@@ -7,7 +7,7 @@ import { Assessment } from '../../../../assessments/types/iassessment';
 import { AssessmentScanPolicyRunner, IIsAnAssessmentSelected, IScheduleScan } from '../../../../background/assessment-scan-policy-runner';
 import { AssessmentStore } from '../../../../background/stores/assessment-store';
 import { VisualizationStore } from '../../../../background/stores/visualization-store';
-import { BaseStore } from '../../../../common/istore';
+import { BaseStore } from '../../../../common/base-store';
 import { IAssessmentData, IAssessmentStoreData } from '../../../../common/types/store-data/iassessment-result-data';
 import { IVisualizationStoreData, TestsEnabledState } from '../../../../common/types/store-data/ivisualization-store-data';
 import { VisualizationType } from '../../../../common/types/visualization-type';
